move commands to common package
removes unused classes as well
