+++ lando/common/tests/test_commands.py
@@ -1,4 +1,4 @@
-from lando.commands import WorkflowNames, WorkflowTypes
+from lando.common.commands import WorkflowNames, WorkflowTypes
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
@@ -13,13 +13,13 @@ class TestWorkflowNamess(TestCase):
         self.invalid_job.workflow.workflow_type = 'other'
         self.mock_names = Mock(WORKFLOW='/workflowdir')
 
-    @patch('lando.commands.ZippedWorkflowNames')
+    @patch('lando.common.commands.ZippedWorkflowNames')
     def test_zipped_type(self, mock_zipped_workflow_names):
         names = WorkflowNames(self.zipped_job, self.mock_names)
         self.assertEqual(names._names_target, mock_zipped_workflow_names.return_value)
         mock_zipped_workflow_names.assert_called_with(self.zipped_job.workflow, '/workflowdir', '/workflowdir/someurl')
 
-    @patch('lando.commands.PackedWorkflowNames')
+    @patch('lando.common.commands.PackedWorkflowNames')
     def test_packed_type(self, mock_packed_workflow_names):
         names = WorkflowNames(self.packed_job, self.mock_names)
         self.assertEqual(names._names_target, mock_packed_workflow_names.return_value)
