+++ lando/k8s/tests/test_jobmanager.py
@@ -1,7 +1,8 @@
 from unittest import TestCase
 from unittest.mock import Mock, call, patch
 from lando.k8s.jobmanager import JobManager, JobStepTypes, StageDataConfig, RunWorkflowConfig, \
-    OrganizeOutputConfig, SaveOutputConfig, RecordOutputProjectConfig, WorkflowTypes, Names, Paths
+    OrganizeOutputConfig, SaveOutputConfig, RecordOutputProjectConfig, Names, Paths
+from lando.common.commands import WorkflowTypes
 import json
